Remove pkg/{dmsg, messaging, messaging-discovery}
