+++ Makefile
@@ -65,9 +65,7 @@ test: ## Run tests
 	#${OPTS} go test -race -tags no_ci -cover -timeout=5m ./pkg/...
 	${OPTS} go test ${TEST_OPTS} ./pkg/app/...
 	${OPTS} go test ${TEST_OPTS} ./pkg/cipher/...
-	${OPTS} go test ${TEST_OPTS} ./pkg/dmsg/...
 	${OPTS} go test ${TEST_OPTS} ./pkg/manager/...
-	${OPTS} go test ${TEST_OPTS} ./pkg/messaging-discovery/...
 	${OPTS} go test ${TEST_OPTS} ./pkg/node/...
 	${OPTS} go test ${TEST_OPTS} ./pkg/route-finder/...
 	${OPTS} go test ${TEST_OPTS} ./pkg/router/...
@@ -75,7 +73,6 @@ test: ## Run tests
 	${OPTS} go test ${TEST_OPTS} ./pkg/setup/...
 	${OPTS} go test ${TEST_OPTS} ./pkg/transport/...
 	${OPTS} go test ${TEST_OPTS} ./pkg/transport-discovery/...
-	${OPTS} go test  -tags no_ci -cover -timeout=5m ./pkg/messaging/...
 
 
 install-linters: ## Install linters
