+++ pkg/transport/dmsg/config.go
@@ -1,0 +1,17 @@
+package dmsg
+
+import (
+	"time"
+
+	"github.com/skycoin/dmsg/cipher"
+	"github.com/skycoin/dmsg/disc"
+)
+
+// Config configures dmsg
+type Config struct {
+	PubKey     cipher.PubKey
+	SecKey     cipher.SecKey
+	Discovery  disc.APIClient
+	Retries    int
+	RetryDelay time.Duration
+}
